fix: install `iptables` for k8s provider if not present (#57)
The k8s snap relies on `iptables` being present on the system, which is
some cases is not true.

In this change, the k8s provider is updated to first check if the
iptables command is present, and install it if not.

In a change from the norm, iptables is not then removed, as currently we
don't track state and removing the package if it was previously
installed would be a bad outcome!

Fixes #56