docs: Add SECURITY.md (#63)
Add a SECURITY.md that explains the security release & disclosure policy
for Concierge. This is strongly based on the ones used in other Charm
Tech projects (themselves based on the recommendations of the Canonical
security team).

This promises security updates for all major versions that have had
releases in the previous year. Right now that would be 0.x and 1.x. We
could exclude 0.x, but I feel that if the choice is made to do 0.x
releases rather than 1.x pre-releases then it's reasonable to allow
people to pin to 0.x and therefore that security releases are a
reasonable expectation.