docs: extend the security documentation (#84)
Extends the SECURITY.md doc to include content that would normally be
found in a Security doc in the Explanation section. If Concierge gains a
documentation site, or documentation on Concierge is added to another
site (e.g. Ops) then this content could move there at that time.

Adds information on:

* Cryptography use (particularly links to the actual use by `apt` and
`snap`)
* Hardening (not required)
* Security risks introduced/changed when using Concierge (none)

I've looked for security and hardening information for the various
packages that Concierge uses, to link to it. Unfortunately, many of them
do not (yet) have any available. I've included the ones that do.

The goal is for this doc to be generally useful for charm authors, but
also to comply with
[SEC0030](https://docs.google.com/document/d/12_hj3_xhaYWqKHsmnuRl5wQO3Es4GuHWtgrxMHG21FM/edit?tab=t.0#heading=h.jszeo1fa8onu)
and [Canonical product security documentation
guidance](https://docs.google.com/document/d/1QBVUxfOVaTXEbBUiiWMzYHiAxG79NIGS7YTVX_GNE4Q/edit?tab=t.0#heading=h.ikq8mvhgar8v).


[Preview](https://github.com/tonyandrewmeyer/concierge/blob/extend-security-doc/SECURITY.md)