fix: use a channel in the override test that is likely to stick around (#112)
The `overrides-env` test currently tries to use the environment to
override the default `k8s` channel to be `1.31-classic/beta`. This
channel appears to have been removed at some point, which causes the
test to now fail.

The PR changes the override to `latest/edge`, which seems likely to be
around for a long time (the default is currently "1.32-classic/stable").