feat: add an option to pass --agent-version when bootstrapping (#111)
Adds a new config option in `juju`, `agent-version`. When specified,
this is passed as `--agent-version` when bootstrapping Juju.

This is particularly useful when there is an issue with the latest agent
version and users want to restrict the agent version to one without the
issue, but otherwise use the latest Juju.